+++ client/src/components/DisplayAllContainer/DisplayAllContainer.js
@@ -1,7 +1,7 @@
 import Container from "react-bootstrap/Container"
 import Col from "react-bootstrap/Col"
 import Row from "react-bootstrap/Row"
-import DisplayCard from "./DisplayCard"
+import DisplayCard from "../DisplayCard/DisplayCard"
 
 function DisplayAllContainer({writing, audio, art, video}) {
