+++ client/src/components/DisplayTypeContainer/DisplayTypeContainer.js
@@ -1,4 +1,4 @@
-import DisplayCard from "./DisplayCard";
+import DisplayCard from "../DisplayCard/DisplayCard";
 import Container from "react-bootstrap/Container"
 import Row from "react-bootstrap/Row"
 import Col from 'react-bootstrap/Col'
