+++ client/src/pages/EditAudioPage/EditAudioPage.js
@@ -8,7 +8,7 @@ import Modal from 'react-bootstrap/Modal'
 import {useSelector} from "react-redux"
 import {useNavigate} from 'react-router-dom'
 import {useEffect, useState} from "react"
-import TopNav from "../navigation/TopNav"
+import TopNav from "../../components/TopNav/TopNav"
 
 function EditAudio() {
     const audio = useSelector(state => state.editCreations.editAudio)
