+++ client/src/pages/MyLibrary/MyLibrary.js
@@ -4,10 +4,10 @@ import Row from 'react-bootstrap/Row'
 import Col from 'react-bootstrap/Col'
 import {useSelector, useDispatch} from "react-redux"
 import {setLibAll, setLibArt, setLibVid, setLibAud, setLibWrit} from "../../state/myLibrarySlice"
-import TopNav from "../navigation/TopNav"
-import DisplayAllContainer from "../display/DisplayAllContainer"
-import DisplayTypeContainer from "../display/DisplayTypeContainer"
-import Sidebar from "../navigation/Sidebar"
+import TopNav from "../../components/TopNav/TopNav"
+import DisplayAllContainer from "../../components/DisplayAllContainer/DisplayAllContainer"
+import DisplayTypeContainer from "../../components/DisplayTypeContainer/DisplayTypeContainer"
+import Sidebar from "../../components/Sidebar/Sidebar"
 import styled from 'styled-components'
 
 function MyLibrary() {
